Removed useless configs
