+++ src/components/EditorResource.js
@@ -6,7 +6,6 @@ const classNames = require('classnames')
 const partial = require('lodash/partial')
 const { withState } = require('recompose')
 const { EditorSchema } = require('./EditorSchema')
-const config = require('../config')
 
 // Pure components
 
@@ -246,7 +245,7 @@ const mapDispatchToProps = (dispatch, { resourceIndex, descriptor }) => ({
     if (descriptor.path.startsWith('http')) {
       dispatch(async () => {
         const table = await Table.load(descriptor.path)
-        const rows = await table.read({ limit: config.EDITOR_UPLOAD_ROWS_LIMIT })
+        const rows = await table.read({ limit: EDITOR_UPLOAD_ROWS_LIMIT })
         const headers = table.headers
         dispatch({
           type: 'UPLOAD_DATA',
@@ -276,7 +275,7 @@ const mapDispatchToProps = (dispatch, { resourceIndex, descriptor }) => ({
       stream.push(text)
       stream.push(null)
       const table = await Table.load(stream)
-      const rows = await table.read({ limit: config.EDITOR_UPLOAD_ROWS_LIMIT })
+      const rows = await table.read({ limit: EDITOR_UPLOAD_ROWS_LIMIT })
       const headers = table.headers
       dispatch({
         type: 'UPLOAD_DATA',
@@ -303,6 +302,10 @@ function readFile(file) {
     }
   })
 }
+
+// Settings
+
+const EDITOR_UPLOAD_ROWS_LIMIT = 100
 
 // Wrappers
